refactor: create the sky box view matrix in the Sky Box's draw method instead of the Renderer class
